refactor: move contract logic to .contract.js file
